Use separate config file for raw-data upload.
This makes for a clearer separation of this component (which one might
want to remove at some point) from the rest of the export script. And
then we don't touch SPSP config at all, we just use it as it is in the
repo.

Also renamed the R script to make it clearer that this is the part that
triggers the external component Raw-Data-Uploader, which has its own
repo.